requirements.txt
asgiref==3.10.0
coverage==7.11.0
Django==5.0.14
djangorestframework==3.16.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
